Used logic loop to perform tests of CSV Read/Write and program logic. Associated bug fixes, and modification to object tests to allow doubles as Item parameters

+++ project/src/delivery/Truck.java
@@ -59,7 +59,7 @@ public abstract class Truck {
 	 * @throws DeliveryException Exception thrown by Truck types with no set temperature.
 	 * @author Tim
 	 */
-	abstract public int getTemp() throws DeliveryException;
+	abstract public double getTemp() throws DeliveryException;
 
 	
 	/**
